Skapade filerna model.dart och ui.dart
